start screen, store hi scores
