+++ lib/superGhost.js
@@ -4,6 +4,8 @@ function SuperGhost(options){
   this.x = 10;
   this.y = 5;
   this.ctx = options.ctx;
+  this.gameEnd = false;
+
 }
 
 SuperGhost.prototype.chasePlayer = function(playerX, playerY, map){
@@ -14,8 +16,10 @@ SuperGhost.prototype.chasePlayer = function(playerX, playerY, map){
   this.x < playerX ? (this.x += newX) : (this.x -= newX);
   this.y < playerY ? (this.y += newY) : (this.y -= newY);
 
+
   if (Math.floor(this.x) === Math.floor(playerX) && Math.floor(this.y) === Math.floor(playerY)) {
-    console.log("You Lose!");
+    this.gameEnd = true;
+    //old game end
   }
 }
